// app/layout/models/store-informtion.model.ts
export class StoreInfo {
    storeLocationID: number;
    posSystemCD: string;
    tankMonitoringSystemID: number;
    companyID: number;
    paymentSystemsProductCode: number;
    posStoreLocationID: number;
    isAutoPushToPOS: boolean;
    isUploadToPOSEnabled: boolean;
    isMixMatchEnabled: boolean;
    storeFullName: string;
    storeName: string;
    storeAddressLine1: string;
    storeAddressLine2: string;
    city: string;
    countyCode: string;
    stateCode: string;
    phoneNo: string;
    fax: string;
    e_Mail: string;
    macAddress: string;
    downloadFilesLocation: string;
    storeLocationIdentifier: string;
    storeContactName: string;
    notificationE_Mail: string;
    notificationSMSPhoneNo: string;
    lastPingFromClientDateTime: string;
    softwareVersion: string;
    isInvoiceEnabled: boolean;
    isInPOSSyncStatus: boolean;
    lotteryStateCode: string;
    lotteryRetailerID: number;
    lotteryRetailerUserName: string;
    lotteryRetailerPassword: string;
    otherAppMacAddress: string;
    invoiceAppMacAddress: string;
    invoiceAppUser: string;
    invoiceAppPwd: string;
    isLotteryStartingFromZero: boolean;
    isFuelPriceUpdateEnabled: boolean;
    noOfLotteryBins: number;
    isDayCloseLocation: boolean;
    isJobberOnly: boolean;
    zipCode: string;
    enableDayReconSubmission: boolean;
    isCOGEnabled: boolean;
    posSystemSubCD: number;
    roleName: string;
    stateName: string;
    isSmsSent: boolean;
    smsPayloadTypeID: number;
    smsPayLoadTypeName: string;
    ignoreTotalFileStatusForCompany: boolean;
    ignoreTotalFileStatusForCompanyTillDate: Date;
    ignoreTotalFileStatusForAdmin: boolean;
    ignoreTotalFileStatusForAdminTillDate: Date;
    ignoreTotalFileStatus: boolean;
    ignoreTotalFileStatusTillDate: Date;
    countyName: string;
    paymentSystemsProductName: string;
    pOsSystemName: string;
    tankMonitoringSystemName: string;
    stateRetailerWebsite: string;
    syncUser: string;
    syncPwd: string;
}